Fix onnx warnings in verify and parse tests

+++ test/onnx/gen_onnx.py
@@ -10019,7 +10019,7 @@ def softmaxcrossentropyloss_2d_mean_reduction_half_test():
 def softmaxcrossentropyloss_2d_no_reduction_weighted_test():
     scores = helper.make_tensor_value_info('0', TensorProto.FLOAT, [4, 4])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4])
-    weights = helper.make_tensor_value_info('2', TensorProto.DOUBLE, [4])
+    weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
     loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
@@ -10082,7 +10082,7 @@ def softmaxcrossentropyloss_2d_no_reduction_half_weighted_test():
 def softmaxcrossentropyloss_2d_sum_reduction_weighted_test():
     scores = helper.make_tensor_value_info('0', TensorProto.FLOAT, [4, 4])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4])
-    weights = helper.make_tensor_value_info('2', TensorProto.DOUBLE, [4])
+    weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
     loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
@@ -10145,7 +10145,7 @@ def softmaxcrossentropyloss_2d_sum_reduction_half_weighted_test():
 def softmaxcrossentropyloss_2d_mean_reduction_weighted_test():
     scores = helper.make_tensor_value_info('0', TensorProto.FLOAT, [4, 4])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4])
-    weights = helper.make_tensor_value_info('2', TensorProto.DOUBLE, [4])
+    weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
     loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
@@ -10254,7 +10254,7 @@ def softmaxcrossentropyloss_kd_no_reduction_weighted_test():
                                            [4, 4, 2, 2])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4, 2, 2])
     weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
-    loss = helper.make_tensor_value_info('3', TensorProto.DOUBLE, [4])
+    loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
         "SoftmaxCrossEntropyLoss",
